alert_colletor/requirements/requirements.txt
requests
httpx
python-decouple
attrs
boto3
datetime